+++ CodeStarterCamp_Week1/Bebero.drawio
@@ -1,0 +1,1 @@
+<mxfile host="app.diagrams.net" modified="2022-02-13T07:58:31.583Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/98.0.4758.80 Safari/537.36" etag="9Dyr9FxyRjMw1aIp4bCH" version="16.5.5" type="device"><diagram id="DQ79mKRbKRTrVa9-Sqhf" name="Page-1">7Vtbc6M2FP41esnMdpDE9dFg0p1edjubh3afOsRWbLoYebAc2/31lUAYJIijboLZOH5JpMMREufoOzfJAEer/c9Fsl7+TuckA8ia7wGeAoSgC33+T1AOkmJhXFEWRTqXtIZwl/5LakZJ3aZzslEYGaUZS9cqcUbznMyYQkuKgu5UtgeaqbOukwXpEO5mSdal/pnO2bKi+shr6B9JuljWM0M3qJ7cJ7Nvi4JuczkfQPjWvoXTuHq8Sup3yQ/dLJM53bVIOAY4KihlVWu1j0gmhFuLrRp3+8TT47oLkjOTAb9+gf6UeftfHj5/oR+dNNr97X2QunpMsi2pP6NcLDvUAiJzLi/ZpQVb0gXNkyxuqGEpBCKmsXiv4fmN0jUnQk78hzB2kMpPtoxy0pKtMvk0S+5JFh7lGdGMFuXktURxuGEF/UZaT2LLmzj8W8MHmrMWHQc4clxOrz5ErP5JgUnShm6LGTkhpXpfJsWCsBN86KhWjhdCV4QVBz6uIFnC0kd1HYnct4sjX6M73pDq+x+qhF1VxiEIAxBGSmMCQcBnsEAcgSAGvlMSLTApG6EFghDEExD6oq1vhkbVQm+7ZcrI3TophbfjBkJV60OaZW1doijyb1+gy0dSMLI/rc2u9OsBWH6NNFSu7O4a0MMaqssW4B1rIH1hbwyocWkVh7/E+J+cuvtVvq7sTPdK7yB7PzpEkSFEK+80FkZRD0YrqPklHKcgtAGecBbI+SqUcnCWLPEU+C4Iner5zc2NZEBuxj84vC94ayFakpNDmjcCBwSwGsKfJysB0epv6/0e8APR8CPgW0+tyL5gYwBtQ2tgD2UNgqsjNkCv/0LwyqF/0JTP3OyGQN0NONC0XFkfOUpT9HEZ3697r0f3XVB37IDAKRZOHVk9kOfgDV3h6Y/e/XLhi03hO5gzhz3R1/vD77Ne2jfEObT69X+mSHqUyOxN6O7Vba/tqFC2PUd9RbVXBrO9fk9E1mtp1YhIUCwRLJWRmsibcEnmMRQfbdcmGQquIy++7drgJV3dbzfP21/VWP9Q1tjVNIidjjE+8rSNsT+YMbavxtikqmGZWuNRc6Z6mT3h0Wad5K8ZIFVv5atUXqxvnbeL1G4NpAvV/rjJGUq53bD0ikwJuHPlQ5OiSA4thrXwtZunXba2iTys1Z1Ps9cOvtk11fyv6tVhT13bxK2LvR8Kn37BoPegYbLkDuWfUVe8V9B/gK4h6NG4yZHbDy2thnjB1QdbD3hNqw/eYAFvN4v5Gt91RM4/malyVeWX05xowpakJEsXOe/OuIgIp4dCgOksySbywSqdz0uY9ilShe6AmsF6Mom7munLRPBgiumWdT99fn96OR72S704PYg5q17Q9dzbxCUh0wyxEt9op2rXw5PX1eYoaYVnackpPJ1X6PzQUhORYTIL1FONMDlCbRUM7XdeL8TacZvjdb3BeQuGyLkaEBMDgkwNiD2qO0BGEH02Q3m7CNMzFM80Qxku5XfHAFRz2Qm0rzo1N5/e6mUn2xSILy0VfF+B0Fa3X/CMJ9f5Nc8/kCPvOSQyceSXWCLU7YXvj24vundi3mPi7OjI6FHMeRPnbqUp/jS9YChAyxobC9gf13de1kVhbHq25o0Zw+L+A6zmNv9Epo/t2/w8jA2cJrlEl12G18+10IB3eHm3+aFOFYE0P4fC8X8=</diagram></mxfile>
